+++ lib/kettle/family/workflow.rb
@@ -1065,7 +1065,7 @@ module Kettle
       end
 
       def release_waves(release_members)
-        ReleaseWaves.new(members: release_members).waves
+        ReleaseWaves.new(members: release_members, configured_waves: config.release_waves).waves
       end
 
       def append_dependency_floor_results(released_members:, dependent_members:, runner:, memo:)